fixed recipe deletion & docs

+++ src/services/recipes.js
@@ -3,6 +3,7 @@ import calculatePaginationData from '../utils/calculatePaginationData.js';
 import { normalizeRecipe, normalizeRecipeArray } from '../utils/normalizeRecipeFunc.js';
 import { recipesCollection } from '../models/recipe.js';
 import { userModel } from '../models/user.js';
+import mongoose from 'mongoose';
 
 export const getRecipes = async (params) => {
   const { page, perPage, categories = [], ingredients = [], searchQuery = '' } = params;
@@ -54,21 +55,15 @@ export async function createRecipe(payload) {
 }
 
 export async function deleteRecipe(recipeId, userId) {
-  const deletedRecipe = await recipesCollection.findOneAndDelete({
-    _id: recipeId,
-    owner: userId,
+  const session = await mongoose.startSession();
+
+  await session.withTransaction(async () => {
+    const deletedRecipe = await recipesCollection.findOneAndDelete({ _id: recipeId, owner: userId }, { session });
+    if (!deletedRecipe) throw createHttpError(404, 'Recipe not found or you are not the owner');
+    await userModel.updateMany({ favourites: recipeId }, { $pull: { favourites: recipeId } }, { session });
   });
 
-  if (!deletedRecipe) {
-    throw createHttpError(404, "Recipe not found or you are not the owner");
-  }
-
-  await userModel.updateMany(
-    { favourites: recipeId },
-    { $pull: { favourites: recipeId } }
-  );
-
-  return deletedRecipe;
+  session.endSession();
 }
 
 export async function getOwnRecipes(userId) {
@@ -84,33 +79,29 @@ export async function addToFavourites(recipeId, userId) {
   const user = await userModel.findByIdAndUpdate(
     userId,
     { $addToSet: { favourites: recipeId } },
-    { new: true, fields: { favourites: 1, _id: 0 } }
+    { new: true, fields: { favourites: 1, _id: 0 } },
   );
-  if (!user) throw createHttpError(404, "User not found");
+  if (!user) throw createHttpError(404, 'User not found');
   const recipe = await recipesCollection
     .findById(recipeId)
     .populate({ path: 'ingredients.id', select: '-_id' })
     .lean()
     .exec();
-    return normalizeRecipe(recipe);
+  return normalizeRecipe(recipe);
 }
 
 export async function removeFromFavourites(recipeId, userId) {
-  await userModel.findByIdAndUpdate(
-    userId,
-    { $pull: { favourites: recipeId } },
-    { new: true }
-  );
+  await userModel.findByIdAndUpdate(userId, { $pull: { favourites: recipeId } }, { new: true });
   const recipe = await recipesCollection
     .findById(recipeId)
     .populate({ path: 'ingredients.id', select: '-_id' })
     .lean()
     .exec();
-    return normalizeRecipe(recipe);
+  return normalizeRecipe(recipe);
 }
 
 export async function getFavouriteRecipes(userId) {
-  const {favourites} = await userModel
+  const { favourites } = await userModel
     .findById(userId)
     .populate({
       path: 'favourites',
@@ -122,6 +113,5 @@ export async function getFavouriteRecipes(userId) {
     })
     .lean()
     .exec();
-    console.log(favourites)
   return { items: normalizeRecipeArray(favourites) };
 }
